+++ docs/history.rst
@@ -5,8 +5,8 @@ Release History
 
 `Next Release`_
 ---------------
-- Add :class:`sprockets.clients.http.HTTPClient`
-- Add :class:`sprockets.clients.http.ClientMixin`
-- Add :class:`sprockets.clients.http.HTTPError`
+- Add :class:`sprockets.clients.http.client.HTTPClient`
+- Add :class:`sprockets.clients.http.mixins.ClientMixin`
+- Add :class:`sprockets.clients.http.client.HTTPError`
 
 .. _Next Release: https://github.com/sprockets/sprockets.clients.http/compare/0.0.0...master